Crate heuristic_common

AHC共通ライブラリ

ヒューリスティックコンテスト用の共通機能を提供するライブラリです。

主な機能

グリッド操作: 2D座標とグリッドの操作

ユーティリティ: ランダム関数、タイマー、デバッグ支援

アルゴリズム: よく使用されるアルゴリズムの実装

入出力: 標準的な入出力処理

ビジュアライザー: デバッグ用データ出力支援

使用例

use heuristic_common::*;
 
fn main() {
    // グリッドの作成
    let mut grid = Grid::new(10, 10, 0);
    let pos = Pos::new(5, 5);
    grid.set(pos, 42);
     
    // タイマーの使用
    let timer = Timer::new(2000); // 2秒制限
     
    // デバッグ出力
    debug!("Current position: {:?}", pos);
     
    // 時間測定
    let result = measure_time!("calculation", {
        // 重い計算処理
        42
    });
}

モジュール構成

Re-exports

pub use utils::*;

pub use algorithms::*;

pub use io::*;

Modules

algorithms

アルゴリズム実装

io

入出力関連ユーティリティ

utils

ユーティリティ関数

visualization

ビジュアライザー支援機能

Macros

debug

デバッグビルド時のみ出力されるデバッグマクロ

measure_time

実行時間を測定するマクロ

Structs

StdRng

The standard RNG. The PRNG algorithm in StdRng is chosen to be efficient on the current platform, to be statistically strong and unpredictable (meaning a cryptographically secure PRNG).

ThreadRng

A reference to the thread-local generator

Traits

CryptoRng

A marker trait used to indicate that an RngCore or BlockRngCore implementation is supposed to be cryptographically secure.

Distribution

Types (distributions) that can be used to create a random instance of T.

IteratorRandom

Extension trait on iterators, providing random sampling methods.

Itertools

An Iterator blanket implementation that provides extra adaptors and methods.

Rng

An automatically-implemented extension trait on RngCore providing high-level generic methods for sampling values and other convenience methods.

RngCore

The core of a random number generator.

SeedableRng

A random number generator that can be explicitly seeded.

SliceRandom

Extension trait on slices, providing random mutation and sampling methods.

Functions

random

Generates a random value using the thread-local random number generator.

thread_rng

Retrieve the lazily-initialized thread-local random number generator, seeded by the system. Intended to be used in method chaining style, e.g. thread_rng().gen::<i32>(), or cached locally, e.g. let mut rng = thread_rng();. Invoked by the Default trait, making ThreadRng::default() equivalent.

Type Aliases

FxHashMap

Type alias for a hashmap using the fx hash algorithm.

FxHashSet

Type alias for a hashmap using the fx hash algorithm.

HashMap

共通的に使用される型エイリアス

HashSet